+++ src/plugin/js/consts.js
@@ -1,0 +1,9 @@
+export default {
+	lightClass : `has-light-background`,
+	darkClass : `has-dark-background`,
+	themeColorVar : `--c-theme-color`,
+	vertiColorVar : `--v-color`,
+	forceColorVar : `--v-forcecolor`,
+	activeclass : 'active',
+	eventNames: ['ready', 'slidechanged']
+}