+++ src/plugin/js/functions/click-bullet.js
@@ -1,0 +1,11 @@
+import { getNodeIndex } from '../helpers';
+
+export const clickBullet = (event, deck) => {
+	if ((event.target).matches('.verticator li a')) {
+		let currIndexh = (deck.getIndices()).h;
+		let currIndexf = (deck.getIndices()).v;
+		event.preventDefault();
+		let i = getNodeIndex(event.target.parentNode);
+		deck.slide(currIndexh, i, currIndexf);
+	}
+}